Implementing Hash Tables/Double Linked List
Creates a hash table of double linked lists that contains athletes with it's information.
Athletes can be added, searched, deleted, and displayed. 
At first the hash function I chose was the division method but there were too many collisions
therefore to minimize the collisions I instead implemented the multiplication method for creating
my hash function. An advantage of this method is that the hash table size 
is not critical. The constant alpha has to be bigger than zero but less 
than 1 therefore I chose the number 0.234262434. The number of 
collisions decreased when comparing to the hash function implementing 
the division method.